Alteração de alert para confirm

+++ client/src/components/userHome/updateProjectCards/UpdateProjectCards.js
@@ -47,18 +47,24 @@ const ProjectCards = () => {
                   const userId = localStorage.getItem("user_id");
                   const projectId = project.id;
 
-                  alert("Deseja remover esse projeto?");
-                  axios.delete(
-                    `http://localhost:5000/user/remove_project/${userId}/${projectId}`
+                  const confirm = window.confirm(
+                    "Deseja remover esse projeto?"
                   );
-                  alert("Projeto removido!");
-                  window.location.reload(false);
+                  if (confirm == true) {
+                    axios.delete(
+                      `http://localhost:5000/user/remove_project/${userId}/${projectId}`
+                    );
+                    alert("Projeto removido!");
+                    window.location.reload(false);
+                  }
                 }}
               >
-                <CardImg1 className='image1'
+                <CardImg1
+                  className="image1"
                   src={require("../../../assets/project-3.png").default}
                 />
-                <CardImg2 className='image2'
+                <CardImg2
+                  className="image2"
                   src={require("../../../assets/project-4.png").default}
                 />
                 <Project key={project.id}>{project.name}</Project>
@@ -79,28 +85,28 @@ const ProjectCards = () => {
 
                 <SkillTag>
                   {project.skill1 != null && project.skill1.length > 0 ? (
-                    <Tag className='tag'>{project.skill1}</Tag>
+                    <Tag className="tag">{project.skill1}</Tag>
                   ) : null}
                   {project.skill2 != null && project.skill2.length > 0 ? (
-                    <Tag className='tag'>{project.skill2}</Tag>
+                    <Tag className="tag">{project.skill2}</Tag>
                   ) : null}
                   {project.skill3 != null && project.skill3.length > 0 ? (
-                    <Tag className='tag'>{project.skill3}</Tag>
+                    <Tag className="tag">{project.skill3}</Tag>
                   ) : null}
                   {project.skill4 != null && project.skill4.length > 0 ? (
-                    <Tag className='tag'>{project.skill4}</Tag>
+                    <Tag className="tag">{project.skill4}</Tag>
                   ) : null}
                   {project.skill5 != null && project.skill5.length > 0 ? (
-                    <Tag className='tag'>{project.skill5}</Tag>
+                    <Tag className="tag">{project.skill5}</Tag>
                   ) : null}
                   {project.skill6 != null && project.skill6.length > 0 ? (
-                    <Tag className='tag'>{project.skill6}</Tag>
+                    <Tag className="tag">{project.skill6}</Tag>
                   ) : null}
                   {project.skill7 != null && project.skill7.length > 0 ? (
-                    <Tag className='tag'>{project.skill7}</Tag>
+                    <Tag className="tag">{project.skill7}</Tag>
                   ) : null}
                   {project.skill8 != null && project.skill8.length > 0 ? (
-                    <Tag className='tag'>{project.skill8}</Tag>
+                    <Tag className="tag">{project.skill8}</Tag>
                   ) : null}
                 </SkillTag>
               </Card>
